gnome stuff
Discovered that the bottles/cartridges xdg-open issue is fixed with
gnome, or the gnome portal stuff maybe. Yet to be seen if this also
fixes when launching from hyprland. Also hopefully no more crashing, did
ssd/hdd tests with usual tools on linux too, no errors there. Perhaps
it's a hyprland issue lol...If so maybe I'll try berrywm again, or even
lose a bunch of time with awesomewm?